requirements.txt
streamlit
requests
pillow
langchain-google-genai
langchain-community
python-dotenv
aiohttp
asyncio
